Revised file
This file describes for all findings presented to Stanford, if they were found via our CARE pipeline using canonical cohorts or curated cohorts. It also describes if a finding was a pan-cancer outlier, pan-disease ouliter, expressed mutation, highly expressed gene (95th percentile). For pd outliers it further describes if it was a consensus outlier found in at least two cohorts or if it was a single-cohort outlier.